+++ agile-interfaces/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  
+  <parent>
+    <groupId>iot.agile</groupId>
+    <artifactId>dbus-api</artifactId>
+    <version>1.0</version>
+  </parent>
+    
+  <name>Agile DBUS Interfaces</name>
+  <artifactId>agile-interfaces</artifactId>
+  <packaging>jar</packaging>
+
+  <repositories>
+    <repository>
+      <id>local-dbus</id>
+      <url>file://${basedir}/../deps</url>
+    </repository>
+  </repositories>    
+
+</project>
