+++ agile-interfaces/src/main/java/iot/agile/agile/interfaces/DeviceManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Dagmawi Neway Mekuria <d.mekuria@create-net.org>.
+ * Copyright 2016 CREATE-NET
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package iot.agile.devicemanager;
+package iot.agile.agile.interfaces;
 
 import java.util.Map;
