+++ iot.agile.protocol.BLE/src/main/resources/simplelogger.properties
@@ -1,0 +1,15 @@
+
+org.slf4j.simpleLogger.defaultLogLevel=DEBUG
+org.slf4j.simpleLogger.showDateTime=true
+org.slf4j.simpleLogger.showShortLogName=true
+
+#org.slf4j.simpleLogger.logFile - The output target which can be the path to a file, or the special values "System.out" and "System.err". Default is "System.err".
+#org.slf4j.simpleLogger.defaultLogLevel - Default log level for all instances of SimpleLogger. Must be one of ("trace", "debug", "info", "warn", or "error"). If not specified, defaults to "info".
+#org.slf4j.simpleLogger.log.a.b.c - Logging detail level for a SimpleLogger instance named "a.b.c". Right-side value must be one of "trace", "debug", "info", "warn", or "error". When a SimpleLogger named "a.b.c" is initialized, its level is assigned from this property. If unspecified, the level of nearest parent logger will be used, and if none is set, then the value specified by org.slf4j.simpleLogger.defaultLogLevel will be used.
+#org.slf4j.simpleLogger.showDateTime - Set to true if you want the current date and time to be included in output messages. Default is false
+#org.slf4j.simpleLogger.dateTimeFormat - The date and time format to be used in the output messages. The pattern describing the date and time format is defined by SimpleDateFormat. If the format is not specified or is invalid, the number of milliseconds since start up will be output.
+#org.slf4j.simpleLogger.showThreadName -Set to true if you want to output the current thread name. Defaults to true.
+#org.slf4j.simpleLogger.showLogName - Set to true if you want the Logger instance name to be included in output messages. Defaults to true.
+#org.slf4j.simpleLogger.showShortLogName - Set to true if you want the last component of the name to be included in output messages. Defaults to false.
+#org.slf4j.simpleLogger.levelInBrackets - Should the level string be output in brackets? Defaults to false.
+#org.slf4j.simpleLogger.warnLevelString - The string value output for the warn level. Defaults to WARN.
